fix: udpate create meeting template logic
